requirements.txt

jira
pandas
yagmail
python-dotenv
openpyxl
